refactor: improve usage of prismic types and relocate slices
