+++ package.json
@@ -41,14 +41,15 @@
 	},
 	"dependencies": {
 		"@prismicio/client": "^6.7.3",
-		"@prismicio/helpers": "^2.3.9"
+		"@prismicio/helpers": "^2.3.9",
+		"@sveltejs/kit": "^1.8.3",
+		"svelte": "^3.55.1"
 	},
 	"devDependencies": {
 		"@prismicio/mock": "^0.2.0",
 		"@prismicio/types": "^0.2.7",
 		"@rollup/plugin-inject": "^5.0.3",
-		"@sveltejs/kit": "^1.8.3",
-		"@sveltejs/package": "^2.0.2",
+		"@sveltejs/package": "^1.0.2",
 		"@typescript-eslint/eslint-plugin": "^5.53.0",
 		"@typescript-eslint/parser": "^5.53.0",
 		"c8": "^7.13.0",
@@ -62,7 +63,6 @@
 		"prettier-plugin-svelte": "^2.9.0",
 		"shx": "^0.3.4",
 		"standard-version": "^9.3.2",
-		"svelte": "^3.55.1",
 		"svelte-check": "^3.0.3",
 		"svelte-preprocess": "^5.0.1",
 		"svelte2tsx": "^0.6.1",
