+++ src/routes/slices/RichText/index.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import * as prismicH from "@prismicio/helpers";
-	import type * as prismicT from "@prismicio/types";
+	import type { RichTextSlice } from "./../../../types/RichTextSlice.type";
 
-	export let slice: prismicT.Slice;
+	export let slice: RichTextSlice;
 </script>
 
 {@html prismicH.asHTML(slice.primary.content)}
