+++ src/routes/slices/index.ts
