+++ src/types/PageDocument.type.ts
@@ -2,10 +2,6 @@ import type * as prismicT from "@prismicio/types";
 import type { CodeSnippetSlice } from "./CodeSnippetSlice.type";
 import type { RichTextSlice } from "./RichTextSlice.type";
 
-export interface PageData extends prismicT.PrismicDocument {
-	document: {
-		data: {
-			body: [CodeSnippetSlice | RichTextSlice];
-		};
-	};
-}
+export type PageDocument = prismicT.PrismicDocument<{
+	body: prismicT.SliceZone<CodeSnippetSlice | RichTextSlice>;
+}>;
