+++ src/types/index.ts
@@ -1,5 +1,5 @@
 export type { CodeSnippetSlice } from "./CodeSnippetSlice.type";
-export type { PageData } from "./PageData.type";
+export type { PageData } from "./PageDocument.type";
 export type { PrismicConfig } from "./PrismicConfig.type";
 export type { PrismicContext } from "./PrismicContext.type";
 export type { RichTextSlice } from "./RichTextSlice.type";
